move rizin and radare to aglib
