+++ pwndbg/commands/__init__.py
@@ -1079,6 +1079,7 @@ def load_commands() -> None:
     import pwndbg.commands.context
     import pwndbg.commands.cpsr
     import pwndbg.commands.cyclic
+    import pwndbg.commands.decompiler_integration
     import pwndbg.commands.dev
     import pwndbg.commands.distance
     import pwndbg.commands.dt
@@ -1092,7 +1093,6 @@ def load_commands() -> None:
     import pwndbg.commands.hex2ptr
     import pwndbg.commands.hexdump
     import pwndbg.commands.hijack_fd
-    import pwndbg.commands.integration
     import pwndbg.commands.jemalloc
     import pwndbg.commands.kbase
     import pwndbg.commands.kbpf
